requirements.txt
Django==4.2.11
djangorestframework==3.14.0
psycopg2-binary==2.9.9
django-redis==5.4.0
python-decouple==3.8
requests==2.31.0
django-cors-headers==4.3.1
django-ratelimit==4.1.0
drf-spectacular==0.27.0
gunicorn==21.2.0
whitenoise==6.6.0
django-environ==0.11.2
dj-database-url==2.1.0
coverage==7.4.0
pytest==8.0.0
pytest-django==4.8.0
factory-boy==3.3.0
freezegun==1.4.0
responses==0.25.0
